pyproject.toml
[build-system]
# build the package with [flit](https://flit.readthedocs.io)
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
# See https://www.python.org/dev/peps/pep-0621/
name = "aiida-bigdft"
dynamic = ["version"]  # read from aiida_bigdft/__init__.py
description = "Translation layer for AiiDA-PyBigDFT"
authors = [{name = "Louis Beal", email = "louis.j.beal@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Development Status :: 3 - Alpha",
    "Framework :: AiiDA"
]
keywords = ["aiida", "plugin"]
requires-python = ">=3.7"
dependencies = [
    "aiida-core>=1.6.3,<3",
    "voluptuous",
    "PyBigDFT",
    "ase",
    "click"
]

[project.urls]
Source = "https://github.com/ljbeal/aiida-bigdft"

[project.optional-dependencies]
testing = [
    "pgtest~=1.3.1",
    "wheel~=0.31",
    "coverage[toml]",
    "pytest~=6.0",
    "pytest-cov"
]
pre-commit = [
    "pre-commit~=2.2",
    "pylint~=2.15.10"
]
docs = [
    "sphinx",
    "sphinxcontrib-contentui",
    "sphinxcontrib-details-directive",
    "furo",
    "markupsafe<2.1"
]

[project.entry-points."aiida.data"]
"bigdft" = "aiida_bigdft.data.BigDFTParameters:BigDFTParameters"
"bigdftfile" = "aiida_bigdft.data.BigDFTFile:BigDFTFile"
"bigdftlogfile" = "aiida_bigdft.data.BigDFTFile:BigDFTLogfile"

[project.entry-points."aiida.calculations"]
"bigdft" = "aiida_bigdft.calculations:BigDFTCalculation"

[project.entry-points."aiida.parsers"]
"bigdft" = "aiida_bigdft.parsers:BigDFTParser"

[project.entry-points."aiida.cmdline.data"]
"bigdft" = "aiida_bigdft.cli:data_cli"

[project.entry-points."aiida.workflows"]
"bigdft" = "aiida_bigdft.workflows.base:BigDFTBaseWorkChain"
"bigdft.relax" = "aiida_bigdft.workflows.relax:BigDFTRelaxWorkChain"

[tool.flit.module]
name = "aiida_bigdft"

[tool.pylint.format]
max-line-length = 125

[tool.pylint.messages_control]
disable = [
    "too-many-ancestors",
    "invalid-name",
    "duplicate-code",
]

[tool.pytest.ini_options]
# Configuration for [pytest](https://docs.pytest.org)
python_files = "test_*.py example_*.py"
filterwarnings = [
    "ignore::DeprecationWarning:aiida:",
    "ignore:Creating AiiDA configuration folder:",
    "ignore::DeprecationWarning:plumpy:",
    "ignore::DeprecationWarning:yaml:",
]

[tool.coverage.run]
# Configuration of [coverage.py](https://coverage.readthedocs.io)
# reporting which lines of your plugin are covered by tests
source=["aiida_bigdft"]

[tool.isort]
# Configuration of [isort](https://isort.readthedocs.io)
line_length = 120
force_sort_within_sections = true
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'AIIDA', 'FIRSTPARTY', 'LOCALFOLDER']
known_aiida = ['aiida']

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py38

[testenv]
usedevelop=True

[testenv:py{37,38,39,310}]
description = Run the test suite against a python version
extras = testing
commands = pytest {posargs}

[testenv:pre-commit]
description = Run the pre-commit checks
extras = pre-commit
commands = pre-commit run {posargs}

[testenv:docs]
description = Build the documentation
extras = docs
commands = sphinx-build -nW --keep-going -b html {posargs} docs/source docs/build/html
commands_post = echo "open file://{toxinidir}/docs/build/html/index.html"
"""
